requirements.txt
google-genai
python-dotenv
tqdm
jinja2
